dma_intel: fix the wrong unformat type
The unformat type for "%d" should be u32 or int, otherwise the 'did' in
high stack address will be overflow to zero by the 'qid' which is in the
low stack address.

Like input "dev wq3.2" will return "did=0, qid=2".

Type: fix

+++ src/plugins/dma_intel/main.c
@@ -222,7 +222,7 @@ dsa_config (vlib_main_t *vm, unformat_input_t *input)
 {
   clib_error_t *error = 0;
   intel_dsa_channel_t *ch;
-  u8 did, qid;
+  u32 did, qid;
 
   if (intel_dsa_main.lock == 0)
     clib_spinlock_init (&(intel_dsa_main.lock));
